requirements.txt
gradio
basicsr
opencv-python
opencv-python-headless
einops
torch
torch-tensorrt
torchtext
torchvision
torchaudio
pyyaml